Switch from Travis to GitHub Actions for CI
Travis stopped building for open source in June 2021.

I'm testing Ruby 2.4+ and activerecord 4.2+, using Rails's CI as a guide
for which Ruby versions to test each activerecord version on. While the
gemspec supports Ruby 1.8+ and activerecord 3.1+, the older Ruby
versions with aren't installable on GitHub Actions, or crash, so they're
not tested. Given their age, support should likely be dropped for them
as compatibility will no longer be ensured if CI isn't testing them.

I'm using PostgreSQL 11 as v12 and newer use a larger precision for the
distance returned when using `selecting_distance_from`, which fails a
spec. Ideally the spec should be updated to not assert a specific
precision, allowing larger precisions.

This'll highlight a regression in `selecting_distance_from` on
activerecord 6.1+ when called directly on a geolocated model, as opposed
to through a geolocated model, which I'll address in a follow up PR.