Only join to `through_table` from `selecting_distance_from` when there is one
Prior to Rails 6.1, the arguments to `joins!` were compacted, so if you
called `selecting_distance_from` directly on the geolocated model where
`through_table` was `nil`, it'd be compacted away and a join wouldn't
actually be attempted.

Since Rails 6.1, the arguments to `joins!` are no longer compacted, so a
join will be attempted to `nil` in such a case, leading to an error.

Now we'll only join to the `through_table` if there actually is one.

Here's the relevant Rails commit:
https://github.com/rails/rails/commit/6038755377533ca8f779cc9158f4a8708de166b0